remove schema and move to task
